requirements.txt
atproto

python-dotenv

requests

beautifulsoup4
